Fix: Use NaCl box for token decryption
The decryption method for access and refresh tokens was incorrectly
using `crypto_box_seal_open`. This commit updates the decryption to use
the regular `crypto_box_open_easy` method, which is the correct way to
decrypt tokens.

This commit also adds more detailed logging during the decryption
process, and a `getTokenHealth` method to the `LocalStorageService`.